fix(authorization): split client-safe roles from server-only helpers
BoardSettingsMenu (a "use client" component) imported ROLE_RANK from
the @/lib/authorization barrel, which re-exported server-only helpers
that ultimately import next/headers via lib/supabase/server. Next.js
rejects pulling next/headers into a client module graph, breaking the
production build.

Extract ROLE_RANK and the Role type into lib/authorization/roles.ts
(no server imports). The barrel still re-exports both, and the client
component now imports them directly from the roles module.
