+++ lib/authorization/index.ts
@@ -1,3 +1,4 @@
-export type { Role } from "./board";
-export { getBoardRole, ROLE_RANK, requireBoardRole } from "./board";
+export { getBoardRole, requireBoardRole } from "./board";
+export type { Role } from "./roles";
+export { ROLE_RANK } from "./roles";
 export { getWorkspaceRole, requireWorkspaceRole } from "./workspace";
